requirements.txt
langchain==0.3.25
langgraph==0.4.3
langchain_google_genai==2.0.10
langchain_core==0.3.59
langchain_community==0.3.24
psycopg2-binary==2.9.10
google-generativeai==0.8.5
python-dotenv==1.1.0
dateparser==1.2.1
pandas==2.2.3
beautifulsoup4==4.13.4